requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
asttokens=2.4.1=pypi_0
attrs=24.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
ca-certificates=2024.7.2=haa95532_0
certifi=2024.7.4=pypi_0
cffi=1.17.0=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.15.4=pypi_0
fsspec=2024.6.1=pypi_0
grpcio=1.66.1=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.4.0=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pickleshare=0.7.5=pypi_0
pip=24.2=py39haa95532_0
pipreqs=0.5.0=pypi_0
platformdirs=4.2.2=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.28.0=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pywin32=306=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.20.0=pypi_0
setuptools=72.1.0=py39haa95532_0
six=1.16.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
sympy=1.13.2=pypi_0
tensorboard=2.17.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tinycss2=1.3.0=pypi_0
torch=2.0.0=pypi_0
torchaudio=2.0.1=pypi_0
torchsummary=1.5.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
vc=14.40=h2eaa2aa_0
vs2015_runtime=14.40.33807=h98bb1dd_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py39haa95532_0
yarg=0.1.9=pypi_0
zipp=3.20.0=pypi_0
